test: Add thumbnails.router.test, various fixes

+++ tests/_internal/with-temp.function.ts
@@ -1,12 +1,15 @@
 export function withTemp(
-  test: (tmpDir: string) => Promise<void>,
+  test: (tmpDir: string) => Promise<void> | void,
+  dontClean?: boolean,
 ): () => Promise<void> {
   return async () => {
     const tmpDir = await Deno.makeTempDir();
     try {
       await test(tmpDir);
     } finally {
-      await Deno.remove(tmpDir, { recursive: true });
+      if (!dontClean) {
+        await Deno.remove(tmpDir, { recursive: true });
+      }
     }
   };
 }
